fix: cesium mocking and basic tests works
