+++ tests/integration/heights/heights.spec.ts
@@ -1,7 +1,8 @@
 /* eslint-disable @typescript-eslint/no-unsafe-return */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+import fs from 'fs';
+import path from 'path';
 import jsLogger from '@map-colonies/js-logger';
-import Cesium, { Cartesian2, Cartographic, sampleTerrainMostDetailed, TerrainProvider } from 'cesium';
+import { Cartesian2 } from 'cesium';
 import { CesiumTerrainProvider } from 'cesium';
 import { Application } from 'express';
 import httpStatusCodes from 'http-status-codes';
@@ -29,20 +30,6 @@ describe('heights', function () {
 
   beforeAll(async function () {
     cesiumTerrainProviderFromUrlSpy = jest.spyOn(CesiumTerrainProvider, 'fromUrl');
-    
-    //@ts-ignore
-    jest.mock('cesium', () => {
-      const originalCesium = jest.requireActual("cesium");
-
-      return {
-        // eslint-disable-next-line @typescript-eslint/naming-convention
-        __esModule: true,
-        ...originalCesium,
-        sampleTerrainMostDetailed: jest.fn().mockImplementation(async (provider: CesiumTerrainProvider, positions: Cesium.Cartographic[]): Promise<Cartographic[]> => {
-          return Promise.resolve(positions.map(pos => ({...pos, height: 66}) as Cartographic));
-        })
-      }
-    });
 
     cesiumTerrainProviderFromUrlSpy.mockReturnValue({
       availability: {
@@ -70,7 +57,7 @@ describe('heights', function () {
   });
 
   describe('Given valid params', function () {
-    describe('Get points height', function () {
+    describe('Get points height (JSON)', function () {
       it('should return 200 status code and points heights for basic usage',async function () {
         const response = await requestSender.getPoints(mockJsonData);
         
@@ -101,7 +88,19 @@ describe('heights', function () {
         }
       });
     });
-  });
 
-  // TODO: Protobuf tests.
+    describe('Get points height (PROTO)', function () {
+      it('should return 200 status code and points heights for basic usage',async function () {
+        const MOCK_PROTO_FILE_RELATIVE_PATH = '../../../src/heights/MOCKS/protoReq.bin';
+        const protoFile = fs.readFileSync(path.resolve(__dirname, MOCK_PROTO_FILE_RELATIVE_PATH), null);
+
+        const response = await requestSender.getPointsProtobuf(protoFile);
+        
+        expect(response.status).toBe(httpStatusCodes.OK);
+        expect(response.type).toBe('application/octet-stream');
+
+        // TODO: should decode returned protobuf data and check as json?
+      });
+    });
+  });
 });
